# Optimisations for Linux GCC: Don't export symbols from DSOs by default - see http://gcc.gnu.org/wiki/Visibility
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@4812 3db68df2-c116-0410-a063-a993310a9797

+++ source/collada/PMDConvert.cpp
@@ -21,7 +21,7 @@
 #include <cassert>
 #include <vector>
 
-const int maxInfluences = 4;
+const size_t maxInfluences = 4;
 struct VertexBlend
 {
 	uint8 bones[maxInfluences];
